refactor code, extract into multiple files
