+++ code/lib/cli/src/automigrate/index.ts
@@ -7,20 +7,14 @@ import tempy from 'tempy';
 import dedent from 'ts-dedent';
 
 import { join } from 'path';
-import {
-  frameworkPackages,
-  getStorybookInfo,
-  loadMainConfig,
-  rendererPackages,
-} from '@storybook/core-common';
+import { getStorybookInfo, loadMainConfig } from '@storybook/core-common';
 import semver from 'semver';
 import { JsPackageManagerFactory, useNpmWarning } from '../js-package-manager';
-import type { PackageManagerName } from '../js-package-manager';
 
-import type { Fix } from './fixes';
-import { allFixes } from './fixes';
+import type { Fix, FixId, FixOptions, FixSummary } from './fixes';
+import { FixStatus, PreCheckFailure, allFixes } from './fixes';
 import { cleanLog } from './helpers/cleanLog';
-import type { InstallationMetadata } from '../js-package-manager/parsePackageInfo';
+import { getMigrationSummary } from './helpers/getMigrationSummary';
 
 const logger = console;
 const LOG_FILE_NAME = 'migration-storybook.log';
@@ -46,44 +40,6 @@ const augmentLogsToFile = () => {
 const cleanup = () => {
   process.stdout.write = originalStdOutWrite;
   process.stderr.write = originalStdErrWrite;
-};
-
-type FixId = string;
-
-interface FixOptions {
-  fixId?: FixId;
-  list?: boolean;
-  fixes?: Fix[];
-  yes?: boolean;
-  dryRun?: boolean;
-  useNpm?: boolean;
-  packageManager?: PackageManagerName;
-  configDir?: string;
-  renderer?: string;
-  skipInstall?: boolean;
-}
-
-enum PreCheckFailure {
-  UNDETECTED_SB_VERSION = 'undetected_sb_version',
-  MAINJS_NOT_FOUND = 'mainjs_not_found',
-  MAINJS_EVALUATION = 'mainjs_evaluation_error',
-}
-
-enum FixStatus {
-  CHECK_FAILED = 'check_failed',
-  UNNECESSARY = 'unnecessary',
-  MANUAL_SUCCEEDED = 'manual_succeeded',
-  MANUAL_SKIPPED = 'manual_skipped',
-  SKIPPED = 'skipped',
-  SUCCEEDED = 'succeeded',
-  FAILED = 'failed',
-}
-
-type FixSummary = {
-  skipped: FixId[];
-  manual: FixId[];
-  succeeded: FixId[];
-  failed: Record<FixId, string>;
 };
 
 const logAvailableMigrations = () => {
@@ -318,184 +274,3 @@ export const automigrate = async ({
 
   return fixResults;
 };
-
-function getMigrationSummary({
-  fixResults,
-  fixSummary,
-  logFile,
-  installationMetadata: repoMetadata,
-}: {
-  fixResults: Record<string, FixStatus>;
-  fixSummary: FixSummary;
-  installationMetadata: InstallationMetadata;
-  logFile?: string;
-}) {
-  const hasNoFixes = Object.values(fixResults).every((r) => r === FixStatus.UNNECESSARY);
-  const hasFailures = Object.values(fixResults).some(
-    (r) => r === FixStatus.FAILED || r === FixStatus.CHECK_FAILED
-  );
-  // eslint-disable-next-line no-nested-ternary
-  const title = hasNoFixes
-    ? 'No migrations were applicable to your project'
-    : hasFailures
-    ? 'Migration check ran with failures'
-    : 'Migration check ran successfully';
-
-  const successfulFixesMessage =
-    fixSummary.succeeded.length > 0
-      ? `
-      ${chalk.bold('Successful migrations:')}\n\n ${fixSummary.succeeded
-          .map((m) => chalk.green(m))
-          .join(', ')}
-    `
-      : '';
-
-  const failedFixesMessage =
-    Object.keys(fixSummary.failed).length > 0
-      ? `
-    ${chalk.bold('Failed migrations:')}\n ${Object.entries(fixSummary.failed).reduce(
-          (acc, [id, error]) => {
-            return `${acc}\n${chalk.redBright(id)}:\n${error}\n`;
-          },
-          ''
-        )}
-    \nYou can find the full logs in ${chalk.cyan(logFile)}\n`
-      : '';
-
-  const manualFixesMessage =
-    fixSummary.manual.length > 0
-      ? `
-      ${chalk.bold('Manual migrations:')}\n\n ${fixSummary.manual
-          .map((m) =>
-            fixResults[m] === FixStatus.MANUAL_SUCCEEDED ? chalk.green(m) : chalk.blue(m)
-          )
-          .join(', ')}
-    `
-      : '';
-
-  const skippedFixesMessage =
-    fixSummary.skipped.length > 0
-      ? `
-      ${chalk.bold('Skipped migrations:')}\n\n ${fixSummary.skipped
-          .map((m) => chalk.cyan(m))
-          .join(', ')}
-    `
-      : '';
-
-  const divider = '\n─────────────────────────────────────────────────\n\n';
-
-  let summaryMessage = dedent`
-    ${successfulFixesMessage}${manualFixesMessage}${failedFixesMessage}${skippedFixesMessage}${
-    hasNoFixes ? '' : divider
-  }If you'd like to run the migrations again, you can do so by running '${chalk.cyan(
-    'npx storybook@next automigrate'
-  )}'
-    
-    The automigrations try to migrate common patterns in your project, but might not contain everything needed to migrate to the latest version of Storybook.
-    
-    Please check the changelog and migration guide for manual migrations and more information: ${chalk.yellow(
-      'https://storybook.js.org/migration-guides/7.0'
-    )}
-    And reach out on Discord if you need help: ${chalk.yellow('https://discord.gg/storybook')}
-  `;
-
-  // these packages are aliased by Storybook, so it doesn't matter if they're duplicated
-  const allowList = [
-    '@storybook/csf',
-    // see this file for more info: code/lib/preview/src/globals/types.ts
-    '@storybook/addons',
-    '@storybook/channel-postmessage',
-    '@storybook/channel-websocket',
-    '@storybook/channels',
-    '@storybook/client-api',
-    '@storybook/client-logger',
-    '@storybook/core-client',
-    '@storybook/core-events',
-    '@storybook/preview-web',
-    '@storybook/preview-api',
-    '@storybook/store',
-
-    // see this file for more info: code/ui/manager/src/globals/types.ts
-    '@storybook/components',
-    '@storybook/router',
-    '@storybook/theming',
-    '@storybook/api',
-    '@storybook/manager-api',
-  ];
-
-  const disallowList = [
-    Object.keys(rendererPackages),
-    Object.keys(frameworkPackages),
-    '@storybook/instrumenter',
-  ];
-
-  if (
-    repoMetadata?.duplicatedDependencies &&
-    Object.keys(repoMetadata.duplicatedDependencies).length > 0
-  ) {
-    let majorVersionClashMessage = '';
-    const duplicatedDependenciesMessage = Object.entries(
-      repoMetadata.duplicatedDependencies
-    ).reduce((acc, [dep, versions]) => {
-      if (allowList.includes(dep)) {
-        return acc;
-      }
-
-      const hasMultipleMajorVersions = hasMultipleVersions(versions);
-
-      if (hasMultipleMajorVersions && majorVersionClashMessage === '') {
-        majorVersionClashMessage = chalk.bold(
-          ' Given that some of these versions differ in the major range, there is a higher chance that Storybook will break.'
-        );
-      }
-
-      if (hasMultipleMajorVersions || disallowList.includes(dep)) {
-        return `${acc}\n${chalk.redBright(dep)}:\n${versions.join(', ')}\n`;
-      }
-      return acc;
-    }, '');
-
-    summaryMessage = dedent`
-     ${summaryMessage}
-     ${divider}${chalk.bold(
-      'Attention:'
-    )} You have duplicated dependencies in your project, which can cause unexpected behavior when using Storybook.${majorVersionClashMessage}
-     ${duplicatedDependenciesMessage}
-     
-     You can find more information for a given dependency by running '${chalk.cyan(
-       `${repoMetadata.infoCommand} <package-name>`
-     )}'
-    `;
-  }
-
-  return boxen(summaryMessage, {
-    borderStyle: 'round',
-    padding: 1,
-    title,
-    borderColor: hasFailures ? 'red' : 'green',
-  });
-}
-
-function hasMultipleVersions(versions: string[]) {
-  return versions.find((v) => {
-    const major = semver.major(v);
-    // If major version === 0, treat minor or patch as major
-    if (major === 0) {
-      const minor = semver.minor(v);
-      if (minor === 0) {
-        const patch = semver.patch(v);
-        return versions.some((v2) => {
-          return semver.patch(v2) !== patch;
-        });
-      }
-
-      return versions.some((v2) => {
-        return semver.minor(v2) !== minor;
-      });
-    }
-
-    return versions.some((v2) => {
-      return semver.major(v2) !== major;
-    });
-  });
-}
