+++ code/lib/cli/src/automigrate/types.ts
@@ -1,4 +1,4 @@
-import type { JsPackageManager } from '../js-package-manager';
+import type { JsPackageManager, PackageManagerName } from '../js-package-manager';
 
 export interface CheckOptions {
   packageManager: JsPackageManager;
@@ -21,3 +21,41 @@ export interface Fix<ResultType = any> {
   prompt: (result: ResultType) => string;
   run?: (options: RunOptions<ResultType>) => Promise<void>;
 }
+
+export type FixId = string;
+
+export enum PreCheckFailure {
+  UNDETECTED_SB_VERSION = 'undetected_sb_version',
+  MAINJS_NOT_FOUND = 'mainjs_not_found',
+  MAINJS_EVALUATION = 'mainjs_evaluation_error',
+}
+
+export interface FixOptions {
+  fixId?: FixId;
+  list?: boolean;
+  fixes?: Fix[];
+  yes?: boolean;
+  dryRun?: boolean;
+  useNpm?: boolean;
+  packageManager?: PackageManagerName;
+  configDir?: string;
+  renderer?: string;
+  skipInstall?: boolean;
+}
+
+export enum FixStatus {
+  CHECK_FAILED = 'check_failed',
+  UNNECESSARY = 'unnecessary',
+  MANUAL_SUCCEEDED = 'manual_succeeded',
+  MANUAL_SKIPPED = 'manual_skipped',
+  SKIPPED = 'skipped',
+  SUCCEEDED = 'succeeded',
+  FAILED = 'failed',
+}
+
+export type FixSummary = {
+  skipped: FixId[];
+  manual: FixId[];
+  succeeded: FixId[];
+  failed: Record<FixId, string>;
+};
