+++ code/lib/cli/src/js-package-manager/NPMProxy.ts
@@ -1,6 +1,22 @@
+import sort from 'semver/functions/sort';
 import { JsPackageManager } from './JsPackageManager';
 import type { PackageJson } from './PackageJson';
-import { mapDependenciesNpm } from './parsePackageInfo';
+import type { InstallationMetadata, PackageMetadata } from './types';
+
+type NpmDependency = {
+  version: string;
+  resolved?: string;
+  overridden?: boolean;
+  dependencies?: NpmDependencies;
+};
+
+type NpmDependencies = {
+  [key: string]: NpmDependency;
+};
+
+export type NpmListOutput = {
+  dependencies: NpmDependencies;
+};
 
 export class NPMProxy extends JsPackageManager {
   readonly type = 'npm';
@@ -39,7 +55,7 @@ export class NPMProxy extends JsPackageManager {
 
     try {
       const parsedOutput = JSON.parse(commandResult);
-      return mapDependenciesNpm(parsedOutput);
+      return this.mapDependencies(parsedOutput);
     } catch (e) {
       return undefined;
     }
@@ -94,5 +110,45 @@ export class NPMProxy extends JsPackageManager {
     } catch (e) {
       throw new Error(`Unable to find versions of ${packageName} using npm`);
     }
+  }
+
+  protected mapDependencies(input: NpmListOutput): InstallationMetadata {
+    const acc: Record<string, PackageMetadata[]> = {};
+    const existingVersions: Record<string, string[]> = {};
+    const duplicatedDependencies: Record<string, string[]> = {};
+
+    const recurse = ([name, packageInfo]: [string, NpmDependency]): void => {
+      if (!name || !name.includes('storybook')) return;
+
+      const value = {
+        version: packageInfo.version,
+        location: '',
+      };
+
+      if (!existingVersions[name]?.includes(value.version)) {
+        if (acc[name]) {
+          acc[name].push(value);
+        } else {
+          acc[name] = [value];
+        }
+        existingVersions[name] = sort([...(existingVersions[name] || []), value.version]);
+
+        if (existingVersions[name].length > 1) {
+          duplicatedDependencies[name] = existingVersions[name];
+        }
+      }
+
+      if (packageInfo.dependencies) {
+        Object.entries(packageInfo.dependencies).forEach(recurse);
+      }
+    };
+
+    Object.entries(input.dependencies).forEach(recurse);
+
+    return {
+      dependencies: acc,
+      duplicatedDependencies,
+      infoCommand: 'npm ls --depth=1',
+    };
   }
 }
