+++ code/lib/cli/src/js-package-manager/PNPMProxy.ts
@@ -1,7 +1,26 @@
 import { pathExistsSync } from 'fs-extra';
 import { JsPackageManager } from './JsPackageManager';
 import type { PackageJson } from './PackageJson';
-import { mapDependenciesPnpm } from './parsePackageInfo';
+import type { InstallationMetadata, PackageMetadata } from './types';
+
+type PnpmDependency = {
+  from: string;
+  version: string;
+  resolved: string;
+  dependencies?: PnpmDependencies;
+};
+
+type PnpmDependencies = {
+  [key: string]: PnpmDependency;
+};
+
+type PnpmListItem = {
+  dependencies: PnpmDependencies;
+  peerDependencies: PnpmDependencies;
+  devDependencies: PnpmDependencies;
+};
+
+export type PnpmListOutput = PnpmListItem[];
 
 export class PNPMProxy extends JsPackageManager {
   readonly type = 'pnpm';
@@ -56,7 +75,7 @@ export class PNPMProxy extends JsPackageManager {
 
     try {
       const parsedOutput = JSON.parse(commandResult);
-      return mapDependenciesPnpm(parsedOutput);
+      return this.mapDependencies(parsedOutput);
     } catch (e) {
       return undefined;
     }
@@ -111,5 +130,49 @@ export class PNPMProxy extends JsPackageManager {
     } catch (e) {
       throw new Error(`Unable to find versions of ${packageName} using pnpm`);
     }
+  }
+
+  protected mapDependencies(input: PnpmListOutput): InstallationMetadata {
+    const acc: Record<string, PackageMetadata[]> = {};
+    const existingVersions: Record<string, string[]> = {};
+    const duplicatedDependencies: Record<string, string[]> = {};
+    const items: PnpmDependencies = input.reduce((curr, item) => {
+      const { devDependencies, dependencies, peerDependencies } = item;
+      const allDependencies = { ...devDependencies, ...dependencies, ...peerDependencies };
+      return Object.assign(curr, allDependencies);
+    }, {} as PnpmDependencies);
+
+    const recurse = ([name, packageInfo]: [string, PnpmDependency]): void => {
+      if (!name || !name.includes('storybook')) return;
+
+      const value = {
+        version: packageInfo.version,
+        location: '',
+      };
+
+      if (!existingVersions[name]?.includes(value.version)) {
+        if (acc[name]) {
+          acc[name].push(value);
+        } else {
+          acc[name] = [value];
+        }
+        existingVersions[name] = [...(existingVersions[name] || []), value.version];
+
+        if (existingVersions[name].length > 1) {
+          duplicatedDependencies[name] = existingVersions[name];
+        }
+      }
+
+      if (packageInfo.dependencies) {
+        Object.entries(packageInfo.dependencies).forEach(recurse);
+      }
+    };
+    Object.entries(items).forEach(recurse);
+
+    return {
+      dependencies: acc,
+      duplicatedDependencies,
+      infoCommand: 'pnpm list --depth=1',
+    };
   }
 }
