+++ code/lib/cli/src/js-package-manager/Yarn1Proxy.ts
@@ -1,6 +1,22 @@
 import { JsPackageManager } from './JsPackageManager';
 import type { PackageJson } from './PackageJson';
-import { mapDependenciesYarn1 } from './parsePackageInfo';
+import type { InstallationMetadata, PackageMetadata } from './types';
+import { parsePackageData } from './util';
+
+type Yarn1ListItem = {
+  name: string;
+  children: Yarn1ListItem[];
+};
+
+type Yarn1ListData = {
+  type: 'list';
+  trees: Yarn1ListItem[];
+};
+
+export type Yarn1ListOutput = {
+  type: 'tree';
+  data: Yarn1ListData;
+};
 
 export class Yarn1Proxy extends JsPackageManager {
   readonly type = 'yarn1';
@@ -41,7 +57,7 @@ export class Yarn1Proxy extends JsPackageManager {
 
     try {
       const parsedOutput = JSON.parse(commandResult);
-      return mapDependenciesYarn1(parsedOutput);
+      return this.mapDependencies(parsedOutput);
     } catch (e) {
       return undefined;
     }
@@ -93,5 +109,44 @@ export class Yarn1Proxy extends JsPackageManager {
     } catch (e) {
       throw new Error(`Unable to find versions of ${packageName} using yarn`);
     }
+  }
+
+  protected mapDependencies(input: Yarn1ListOutput): InstallationMetadata {
+    if (input.type === 'tree') {
+      const { trees } = input.data;
+      const acc: Record<string, PackageMetadata[]> = {};
+      const existingVersions: Record<string, string[]> = {};
+      const duplicatedDependencies: Record<string, string[]> = {};
+
+      const recurse = (tree: typeof trees[0]) => {
+        const { children } = tree;
+        const { name, value } = parsePackageData(tree.name);
+        if (!name || !name.includes('storybook')) return;
+        if (!existingVersions[name]?.includes(value.version)) {
+          if (acc[name]) {
+            acc[name].push(value);
+          } else {
+            acc[name] = [value];
+          }
+          existingVersions[name] = [...(existingVersions[name] || []), value.version];
+
+          if (existingVersions[name].length > 1) {
+            duplicatedDependencies[name] = existingVersions[name];
+          }
+        }
+
+        children.forEach(recurse);
+      };
+
+      trees.forEach(recurse);
+
+      return {
+        dependencies: acc,
+        duplicatedDependencies,
+        infoCommand: 'yarn why',
+      };
+    }
+
+    throw new Error('Something went wrong while parsing yarn output');
   }
 }
