+++ code/lib/cli/src/js-package-manager/Yarn2Proxy.ts
@@ -1,6 +1,7 @@
 import { JsPackageManager } from './JsPackageManager';
 import type { PackageJson } from './PackageJson';
-import { mapDependenciesYarn2 } from './parsePackageInfo';
+import type { InstallationMetadata, PackageMetadata } from './types';
+import { parsePackageData } from './util';
 
 // This encompasses both yarn 2 and yarn 3
 export class Yarn2Proxy extends JsPackageManager {
@@ -41,7 +42,7 @@ export class Yarn2Proxy extends JsPackageManager {
     ]);
 
     try {
-      return mapDependenciesYarn2(commandResult);
+      return this.mapDependencies(commandResult);
     } catch (e) {
       return undefined;
     }
@@ -91,5 +92,38 @@ export class Yarn2Proxy extends JsPackageManager {
     } catch (e) {
       throw new Error(`Unable to find versions of ${packageName} using yarn 2`);
     }
+  }
+
+  protected mapDependencies(input: string): InstallationMetadata {
+    const lines = input.split('\n');
+    const acc: Record<string, PackageMetadata[]> = {};
+    const existingVersions: Record<string, string[]> = {};
+    const duplicatedDependencies: Record<string, string[]> = {};
+
+    lines.forEach((packageName) => {
+      if (!packageName || !packageName.includes('storybook')) {
+        return;
+      }
+
+      const { name, value } = parsePackageData(packageName.replaceAll(`"`, ''));
+      if (!existingVersions[name]?.includes(value.version)) {
+        if (acc[name]) {
+          acc[name].push(value);
+        } else {
+          acc[name] = [value];
+        }
+
+        existingVersions[name] = [...(existingVersions[name] || []), value.version];
+        if (existingVersions[name].length > 1) {
+          duplicatedDependencies[name] = existingVersions[name];
+        }
+      }
+    });
+
+    return {
+      dependencies: acc,
+      duplicatedDependencies,
+      infoCommand: 'yarn why',
+    };
   }
 }
