+++ code/lib/cli/src/js-package-manager/types.ts
@@ -1,0 +1,6 @@
+export type PackageMetadata = { version: string; location?: string; reasons?: string[] };
+export type InstallationMetadata = {
+  dependencies: Record<string, PackageMetadata[]>;
+  duplicatedDependencies: Record<string, string[]>;
+  infoCommand: string;
+};
